+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+1.1.0 - (July 13, 2017)
+------------------
+Minor version bump
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
+
 1.0.0 - (June 28, 2017)
 ------------------
 Initial stable release