+++ packages/terra-list/CHANGELOG.md
@@ -3,8 +3,13 @@ Changelog
 
 Unreleased
 ----------
+
+1.1.0 - (July 13, 2017)
+------------------
 ### Added
 * Add preventDefault to Single Select List and MultiSelect List
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
 
 1.0.0 - (June 28, 2017)
 ------------------